Add password valdation to delete item Resurrected error from deletion Delete edit.ejs

+++ controllers/deleteController.js
@@ -1,7 +1,12 @@
 import * as db from '../db/queries.js';
-import { InternalServerError } from '../helper/errors.js';
+import { CustomError, InternalServerError } from '../helper/errors.js';
+
+const pwValidation = pw => pw === 'vaasmontenegro';
 
 export const deleteItem = async (req, res, next) => {
+  if (!pwValidation(req.query.pw)) {
+    return next(new CustomError('Wrong password! Refer to source code', 401));
+  }
   try {
     const { itemId } = req.params;
     await db.deleteItem(Number(itemId));
@@ -13,6 +18,9 @@ export const deleteItem = async (req, res, next) => {
 };
 
 export const deleteCategory = async (req, res, next) => {
+  if (!pwValidation(req.query.pw)) {
+    return next(new CustomError('Wrong password! Refer to source code', 401));
+  }
   try {
     const { categoryId } = req.params;
     await db.deleteCategory(Number(categoryId));
